requirements.txt
pyzmq>=25.0.0
python-dotenv>=1.0.0
uvloop>=0.19.0
pydantic>=2.5.0
pandas>=2.2.0
matplotlib>=3.8.0
rich>=13.7.0
typer>=0.9.0
